Use more accurate variable name for write-errs
The 'dirty' variable name was poorly named because 'dirty' is often used
in the context of temporary files to indicate that the data has changed,
and thus the transaction needs to be committed or such.

The 'dirty' boolean in 'atom' was being used as exactly the opposite,
and indicator that there was an issue meaning the data *should not* be
committed.

This names the variable a little more clearly to indicate that.